Move file to get a matching name
