+++ lib/msf/core/exploit/mixins.rb
@@ -98,4 +98,4 @@ require 'msf/core/exploit/winrm'
 # WebApp
 require 'msf/core/exploit/web'
 
-require 'msf/core/exploit/browser/server'
+require 'msf/core/exploit/remote/server'
